+++ src/main/java/hng_java_boilerplate/config/WebSecurityConfig.java
@@ -83,7 +83,8 @@ public class WebSecurityConfig {
                                         "/swagger-ui/**",
                                         "/api/v1/auth/**",
                                         "/api/v1/waitlist",
-                                        "/api/v1/faqs"
+                                        "/api/v1/faqs",
+                                        "/api/v1/contacts"
                                 ).permitAll()
                                 .requestMatchers("/api/v1/auth/logout", "/api/**").authenticated())
                 .logout(logout -> logout
