--- src/main/java/hng_java_boilerplate/faq/dto/response/CustomResponse.java
@@ -1,4 +1,0 @@
-package hng_java_boilerplate.faq.dto.response;
-
-public record CustomResponse(String status, String message) {
-}
