--- src/main/java/hng_java_boilerplate/faq/dto/response/FaqResponse.java
@@ -1,3 +1,0 @@
-package hng_java_boilerplate.faq.dto.response;
-
-public record FaqResponse(String status, String id, String question, String answer) {}
