+++ src/main/java/hng_java_boilerplate/helpCenter/contactUs/dto/response/CustomResponse.java
@@ -1,0 +1,3 @@
+package hng_java_boilerplate.helpCenter.contactUs.dto.response;
+
+public record CustomResponse(String status, String message) {}
