+++ src/main/java/hng_java_boilerplate/helpCenter/contactUs/entity/Contact.java
@@ -1,0 +1,35 @@
+package hng_java_boilerplate.helpCenter.contactUs.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.util.UUID;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Table(name = "contacts")
+public class Contact {
+    @Id
+    private String id;
+    @Column(nullable = false)
+    private String name;
+    @Column(nullable = false)
+    private String email;
+    @Column(nullable = false)
+    private String phone;
+    @Column(nullable = false)
+    private String message;
+
+    @PrePersist
+    private void prePersist() {
+        if (id == null) {
+            id = UUID.randomUUID().toString();
+        }
+    }
+}
