+++ src/main/java/hng_java_boilerplate/helpCenter/contactUs/repository/ContactUsRepository.java
@@ -1,0 +1,7 @@
+package hng_java_boilerplate.helpCenter.contactUs.repository;
+
+import hng_java_boilerplate.helpCenter.contactUs.entity.Contact;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ContactUsRepository extends JpaRepository<Contact, String> {
+}
