+++ src/main/java/hng_java_boilerplate/helpCenter/faq/dto/response/CustomResponse.java
@@ -1,0 +1,4 @@
+package hng_java_boilerplate.helpCenter.faq.dto.response;
+
+public record CustomResponse(String status, String message) {
+}
