+++ src/main/java/hng_java_boilerplate/helpCenter/faq/dto/response/FaqResponse.java
@@ -1,0 +1,3 @@
+package hng_java_boilerplate.helpCenter.faq.dto.response;
+
+public record FaqResponse(String status, String id, String question, String answer) {}
